breaks project into modules
